@@ -1,29 +1,42 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
+    <PropertyGroup>
+        <TargetFramework>net8.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <LangVersion>latest</LangVersion>
+        <Version>1.8.0</Version>
+        <AssemblyVersion>1.8.0</AssemblyVersion>
+        <FileVersion>1.8.0</FileVersion>
+        <PackageVersion>1.8.0</PackageVersion>
+        <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+        <Authors>SylveonDeko</Authors>
+        <Company>SylveonDeko</Company>
+        <Product>nHentai</Product>
+        <Description>A (full?) nHentai API implementation for .NET</Description>
+        <PackageReadmeFile>README.md</PackageReadmeFile>
+        <PackageLicenseFile>LICENSE</PackageLicenseFile>
+        <PackageProjectUrl>https://github.com/SylveonDeko/NHentaiAPI</PackageProjectUrl>
+        <RepositoryUrl>https://github.com/SylveonDeko/NHentaiAPI.git</RepositoryUrl>
+        <RepositoryType>git</RepositoryType>
+        <PackageTags>nhentai;n-hentai;hentai;api;csharp;dotnet</PackageTags>
+        <GenerateDocumentationFile>true</GenerateDocumentationFile>
+        <NoWarn>$(NoWarn);CS1591</NoWarn>
+        <PackageReleaseNotes>
+            [1.8.0]
+            - Switched to using System.Text.Json instead of Newtonsoft.
+            - Added proper JSON type handling via converters.
+            - Updated to .NET 8.0
+            - Added proper async handling
+            - Updated thumbnail and image urls to work with nhentais' load balancing urls.
+        </PackageReleaseNotes>
+    </PropertyGroup>
 
-  <PropertyGroup>
-    <TargetFrameworks>net5.0;net6.0</TargetFrameworks>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <Authors>Sylveon76, andy840119</Authors>
-    <Product>nHentai</Product>
-    <Description>A (full?) nHentai API implementation for .NET</Description>
-    <Copyright />
-    <PackageProjectUrl>https://github.com/Sylveon76/NHentaiAPI</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/Sylveon76/NHentaiAPI.git</RepositoryUrl>
-    <RepositoryType>git</RepositoryType>
-    <PackageTags>nhentai n-hentai hentai api csharp Sylveon76, andy840119</PackageTags>
-    <PackageReleaseNotes>
-      [1.5.0]
-Refactor api client    
-    </PackageReleaseNotes>
-    <PackageLicenseFile>LICENSE</PackageLicenseFile>
-    <Version>1.7.0</Version>
-  </PropertyGroup>
+    <ItemGroup>
+        <None Include="..\LICENSE" Pack="true" PackagePath=""/>
+        <None Include="..\README.md" Pack="true" PackagePath=""/>
+    </ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
-    <PackageReference Include="FlareSolverrSharp" Version="2.2.6" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="..\LICENSE" Pack="true" PackagePath="" />
-  </ItemGroup>
+    <ItemGroup>
+        <PackageReference Include="System.Text.Json" Version="9.0.0"/>
+    </ItemGroup>
 </Project>
