@@ -1,12 +1,16 @@
-﻿using System.Collections.Generic;
-using Newtonsoft.Json;
+﻿using System.Text.Json.Serialization;
 using NHentaiAPI.Models.Books;
 
-namespace NHentaiAPI.Models.Recommends
+namespace NHentaiAPI.Models.Recommends;
+
+/// <summary>
+///     Represents a list of recommended books
+/// </summary>
+public class BookRecommend
 {
-    public class BookRecommend
-    {
-        [JsonProperty("result")]
-        public List<Book> Result { get; set; }
-    }
+    /// <summary>
+    ///     Gets or sets the list of recommended books
+    /// </summary>
+    [JsonPropertyName("result")]
+    public List<Book> Result { get; set; }
 }
