@@ -1,15 +1,17 @@
-﻿namespace NHentaiAPI.Models.Searches
-{
-    public enum SortBy
-    {
-        /// <summary>
-        /// Default sorting
-        /// </summary>
-        Default,
+﻿namespace NHentaiAPI.Models.Searches;
 
-        /// <summary>
-        /// Sorting by popular
-        /// </summary>
-        Popular
-    }
+/// <summary>
+///     Search sort method
+/// </summary>
+public enum SortBy
+{
+    /// <summary>
+    ///     Default sorting
+    /// </summary>
+    Default,
+
+    /// <summary>
+    ///     Sorting by popular
+    /// </summary>
+    Popular
 }
